DS-508: Move breakpoint keys to its own file since there were circular imports between spacing and breakpoints types
