+++ ngx-fudis/projects/ngx-fudis/src/lib/types/breakpoint-keys.ts
@@ -1,0 +1,6 @@
+/**
+ * Breakpoint keys.
+ *
+ * Used in responsive spacing and grid settings, and calculating breakpoint boundaries
+ */
+export type FudisBreakpointKey = 'default' | 'xs' | 'sm' | 'md' | 'lg' | 'xl' | 'xxl';
