+++ ngx-fudis/projects/ngx-fudis/src/lib/types/types.ts
@@ -1,4 +1,5 @@
 export * from './breakpoints';
+export * from './breakpoint-keys';
 export * from './errorSummary';
 export * from './forms';
 export * from './grid';
